Feature/jenkins warnings (#538)
 * Remove warnings from Jenkins
 * Improve Jenkins behavior

+++ modules/multiresvolume/rendering/histogrammanager.cpp
@@ -71,9 +71,9 @@ bool HistogramManager::buildHistogram(TSP* tsp, unsigned int brickIndex) {
     if (isBstLeaf && isOctreeLeaf) {
         // TSP leaf, read from file and build histogram
         std::vector<float> voxelValues = readValues(tsp, brickIndex);
-        unsigned int numVoxels = voxelValues.size();
+        size_t numVoxels = voxelValues.size();
 
-        for (unsigned int v = 0; v < numVoxels; ++v) {
+        for (size_t v = 0; v < numVoxels; ++v) {
             histogram.add(voxelValues[v], 1.0);
         }
     } else {
@@ -92,8 +92,8 @@ bool HistogramManager::buildHistogram(TSP* tsp, unsigned int brickIndex) {
                 children.push_back(firstChild + c);
             }
         }
-        int numChildren = children.size();
-        for (int c = 0; c < numChildren; c++) {
+        size_t numChildren = children.size();
+        for (size_t c = 0; c < numChildren; c++) {
             // Visit child
             unsigned int childIndex = children[c];
             if (_histograms[childIndex].isValid() || buildHistogram(tsp, childIndex)) {
@@ -167,7 +167,7 @@ bool HistogramManager::saveToFile(const std::string& filename) {
     return false;
     }
 
-    int numHistograms = _histograms.size();
+    size_t numHistograms = _histograms.size();
     file.write(reinterpret_cast<char*>(&numHistograms), sizeof(int));
     file.write(reinterpret_cast<char*>(&_numBins), sizeof(int));
     file.write(reinterpret_cast<char*>(&_minBin), sizeof(float));
@@ -176,9 +176,9 @@ bool HistogramManager::saveToFile(const std::string& filename) {
     int nFloats = numHistograms * _numBins;
     float* histogramData = new float[nFloats];
 
-    for (int i = 0; i < numHistograms; ++i) {
-    int offset = i*_numBins;
-    memcpy(&histogramData[offset], _histograms[i].data(), sizeof(float) * _numBins);
+    for (size_t i = 0; i < numHistograms; ++i) {
+        int offset = i*_numBins;
+        memcpy(&histogramData[offset], _histograms[i].data(), sizeof(float) * _numBins);
     }
 
     file.write(reinterpret_cast<char*>(histogramData), sizeof(float) * nFloats);
